Quick restructuring PlaylistDetailScreen.kt
